Do not presume that jQuery is loaded with the $ global
jQuery can be loaded without the $ global token, to avoid
certain problems with other js packages.

The $ is an alias for jQuery() and jQuery, respectively,
so one can always use on of those instead to make sure
the code is portable.

This patch makes it possible to use a generated story.js
on a website where jquery is already loaded without the
global token.